data_generator/requirements.txt
fastapi==0.115.0
uvicorn==0.32.0
psycopg2-binary==2.9.10
minio==7.2.12
confluent-kafka==2.6.1
Pillow==11.0.0
httpx==0.27.2
numpy==1.26.4
Faker==33.0.0
